+++ src/components/button/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './button.css';
+import './index.css';
 
 export interface ButtonProps {
   /** Is this the principal call to action on the page? */
